requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.2.2
cchardet==2.1.7
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dataclasses==0.8
Django==2.1.4
django-tinymce==3.4.0
django-tinymce4-lite==1.8.0
elasticsearch==7.13.0
fake-useragent==0.1.11
firebase-admin==5.0.1
geographiclib==1.50
geopy==2.1.0
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
googletransx==2.4.2
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
idna-ssl==1.1.0
jsmin==3.0.1
multidict==5.1.0
pandas==1.1.5
proto-plus==1.19.0
protobuf==3.17.3
pycares==4.0.0
pycparser==2.20
pygame==2.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-socks==1.2.4
pytz==2021.1
requests==2.25.1
rsa==4.7.2
schedule==1.1.0
soupsieve==2.2.1
torpy==1.1.6
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.3.0
wincertstore==0.2
yarl==1.6.3
